Refactor tsdb_goto_epoch
The previously supported flags have been simplified to cover two
cases: failing when an epoch doesn't exist and growing an epoch
on demand.

Epochs are always loaded lazily for writes -- they are never
created directly in the call to goto.

Read only mode might not be properly supported at this point.